requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.4
django-better-admin-arrayfield==1.1.0
-e git+https://github.com/misteraverin/django-bulk-admin.git@026b802e34ba2d445552e1fec632ec23872d1e54#egg=django_bulk_admin
django-bulk-update-or-create==0.2.0
django-common-helpers==0.9.2
django-crontab==0.7.1
django-extensions==2.2.9
django-prometheus==2.1.0
django-rest-marshmallow==4.0.2
djangorestframework==3.11.0
google-api-core==1.19.1
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
inapppy==2.4.4
incremental==17.5.0
libsass==0.20.1
marshmallow==3.5.1
msgpack==0.6.2
multidict==4.7.6
oauth2client==3.0.0
packaging==20.4
prometheus-client==0.8.0
protobuf==3.12.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-decouple==3.3
pytz==2019.3
random-username==1.0.2
requests==2.23.0
rsa==4.0
sentry-sdk==0.19.2
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.3.1
uritemplate==3.0.1
urllib3==1.25.9
yarl==1.4.2
zope.interface==5.0.1
